+++ bee-api/bee-rest-api/src/types/dtos.rs
@@ -31,8 +31,6 @@ use bee_protocol::types::peer::Peer;
 use serde::{Deserialize, Serialize, Serializer};
 use serde_json::Value;
 
-use std::convert::{TryFrom, TryInto};
-
 /// The message object that nodes gossip around in the network.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct MessageDto {
@@ -838,7 +836,7 @@ impl From<&Peer> for PeerDto {
                 } else if peer.relation().is_unknown() {
                     RelationDto::Unknown
                 } else {
-                    RelationDto::Discovered
+                    RelationDto::Autopeered
                 }
             },
             connected: peer.is_connected(),
@@ -882,8 +880,8 @@ pub enum RelationDto {
     Known,
     #[serde(rename = "unknown")]
     Unknown,
-    #[serde(rename = "discovered")]
-    Discovered,
+    #[serde(rename = "autopeered")]
+    Autopeered,
 }
 
 /// Describes the heartbeat of a node.