+++ bee-crypto/src/ternary/bigint/i384/mod.rs
@@ -27,7 +27,6 @@ use byteorder::{self, ByteOrder};
 
 use std::{
     cmp::Ordering,
-    convert::{TryFrom, TryInto},
     fmt,
     marker::PhantomData,
     ops::{Deref, DerefMut},