+++ bee-message/tests/milestone_id.rs
@@ -5,7 +5,6 @@ use bee_common::packable::Packable;
 use bee_message::prelude::*;
 
 use core::str::FromStr;
-use std::convert::TryInto;
 
 const MILESTONE_ID: &str = "52fdfc072182654f163f5f0f9a621d729566c74d10037c4d7bbb0407d1e2c649";
 const MILESTONE_ID_INVALID_HEX: &str = "52fdfc072182654f163f5f0f9a621d729566c74d10037c4d7bbb0407d1e2c64x";