+++ bee-api/bee-rest-api/CHANGELOG.md
@@ -19,6 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.1.3 - 2021-11-03
+
+### Removed
+
+- `RelationDto::Discovered` variant;
+
+### Added
+
+- `RelationDto::Autopeered` variant;
+
 ## 0.1.2 - 2021-10-18
 
 ### Added